Crate rusoto_kms

AWS Key Management Service
AWS Key Management Service (AWS KMS) is an encryption and key management web service. This guide describes the AWS KMS operations that you can call programmatically. For general information about AWS KMS, see the AWS Key Management Service Developer Guide .

AWS provides SDKs that consist of libraries and sample code for various programming languages and platforms (Java, Ruby, .Net, macOS, Android, etc.). The SDKs provide a convenient way to create programmatic access to AWS KMS and other AWS services. For example, the SDKs take care of tasks such as signing requests (see below), managing errors, and retrying requests automatically. For more information about the AWS SDKs, including how to download and install them, see Tools for Amazon Web Services.

We recommend that you use the AWS SDKs to make programmatic API calls to AWS KMS.

Clients must support TLS (Transport Layer Security) 1.0. We recommend TLS 1.2. Clients must also support cipher suites with Perfect Forward Secrecy (PFS) such as Ephemeral Diffie-Hellman (DHE) or Elliptic Curve Ephemeral Diffie-Hellman (ECDHE). Most modern systems such as Java 7 and later support these modes.

Signing Requests

Requests must be signed by using an access key ID and a secret access key. We strongly recommend that you do not use your AWS account (root) access key ID and secret key for everyday work with AWS KMS. Instead, use the access key ID and secret access key for an IAM user. You can also use the AWS Security Token Service to generate temporary security credentials that you can use to sign requests.

All AWS KMS operations require Signature Version 4.

Logging API Requests

AWS KMS supports AWS CloudTrail, a service that logs AWS API calls and related events for your AWS account and delivers them to an Amazon S3 bucket that you specify. By using the information collected by CloudTrail, you can determine what requests were made to AWS KMS, who made the request, when it was made, and so on. To learn more about CloudTrail, including how to turn it on and find your log files, see the AWS CloudTrail User Guide.

Additional Resources

For more information about credentials and request signing, see the following:

AWS Security Credentials - This topic provides general information about the types of credentials used for accessing AWS.

Temporary Security Credentials - This section of the IAM User Guide describes how to create and use temporary security credentials.

Signature Version 4 Signing Process - This set of topics walks you through the process of signing a request using an access key ID and a secret access key.

Commonly Used API Operations

Of the API operations discussed in this guide, the following will prove the most useful for most applications. You will likely perform operations other than these, such as creating keys and assigning policies, by using the console.

Encrypt

Decrypt

GenerateDataKey

GenerateDataKeyWithoutPlaintext

If you're using the service, you're probably looking for KmsClient and Kms.

Structs

AliasListEntry
Contains information about an alias.

CancelKeyDeletionRequest

CancelKeyDeletionResponse

ConnectCustomKeyStoreRequest

ConnectCustomKeyStoreResponse

CreateAliasRequest

CreateCustomKeyStoreRequest

CreateCustomKeyStoreResponse

CreateGrantRequest

CreateGrantResponse

CreateKeyRequest

CreateKeyResponse

CustomKeyStoresListEntry
Contains information about each custom key store in the custom key store list.

DecryptRequest

DecryptResponse

DeleteAliasRequest

DeleteCustomKeyStoreRequest

DeleteCustomKeyStoreResponse

DeleteImportedKeyMaterialRequest

DescribeCustomKeyStoresRequest

DescribeCustomKeyStoresResponse

DescribeKeyRequest

DescribeKeyResponse

DisableKeyRequest

DisableKeyRotationRequest

DisconnectCustomKeyStoreRequest

DisconnectCustomKeyStoreResponse

EnableKeyRequest

EnableKeyRotationRequest

EncryptRequest

EncryptResponse

GenerateDataKeyRequest

GenerateDataKeyResponse

GenerateDataKeyWithoutPlaintextRequest

GenerateDataKeyWithoutPlaintextResponse

GenerateRandomRequest

GenerateRandomResponse

GetKeyPolicyRequest

GetKeyPolicyResponse

GetKeyRotationStatusRequest

GetKeyRotationStatusResponse

GetParametersForImportRequest

GetParametersForImportResponse

GrantConstraints

Use this structure to allow cryptographic operations in the grant only when the operation request includes the specified encryption context.

AWS KMS applies the grant constraints only when the grant allows a cryptographic operation that accepts an encryption context as input, such as the following.

Encrypt

Decrypt

GenerateDataKey

GenerateDataKeyWithoutPlaintext

ReEncrypt

AWS KMS does not apply the grant constraints to other operations, such as DescribeKey or ScheduleKeyDeletion.

In a cryptographic operation, the encryption context in the decryption operation must be an exact, case-sensitive match for the keys and values in the encryption context of the encryption operation. Only the order of the pairs can vary.

However, in a grant constraint, the key in each key-value pair is not case sensitive, but the value is case sensitive.

To avoid confusion, do not use multiple encryption context pairs that differ only by case. To require a fully case-sensitive encryption context, use the kms:EncryptionContext: and kms:EncryptionContextKeys conditions in an IAM or key policy. For details, see kms:EncryptionContext: in the AWS Key Management Service Developer Guide .

GrantListEntry
Contains information about an entry in a list of grants.

ImportKeyMaterialRequest

ImportKeyMaterialResponse

KeyListEntry
Contains information about each entry in the key list.

KeyMetadata
Contains metadata about a customer master key (CMK).

This data type is used as a response element for the CreateKey and DescribeKey operations.

KmsClient
A client for the KMS API.

ListAliasesRequest

ListAliasesResponse

ListGrantsRequest

ListGrantsResponse

ListKeyPoliciesRequest

ListKeyPoliciesResponse

ListKeysRequest

ListKeysResponse

ListResourceTagsRequest

ListResourceTagsResponse

ListRetirableGrantsRequest

PutKeyPolicyRequest

ReEncryptRequest

ReEncryptResponse

RetireGrantRequest

RevokeGrantRequest

ScheduleKeyDeletionRequest

ScheduleKeyDeletionResponse

Tag
A key-value pair. A tag consists of a tag key and a tag value. Tag keys and tag values are both required, but tag values can be empty (null) strings.

For information about the rules that apply to tag keys and tag values, see User-Defined Tag Restrictions in the AWS Billing and Cost Management User Guide.

TagResourceRequest

UntagResourceRequest

UpdateAliasRequest

UpdateCustomKeyStoreRequest

UpdateCustomKeyStoreResponse

UpdateKeyDescriptionRequest

Enums

CancelKeyDeletionError
Errors returned by CancelKeyDeletion

ConnectCustomKeyStoreError
Errors returned by ConnectCustomKeyStore

CreateAliasError
Errors returned by CreateAlias

CreateCustomKeyStoreError
Errors returned by CreateCustomKeyStore

CreateGrantError
Errors returned by CreateGrant

CreateKeyError
Errors returned by CreateKey

DecryptError
Errors returned by Decrypt

DeleteAliasError
Errors returned by DeleteAlias

DeleteCustomKeyStoreError
Errors returned by DeleteCustomKeyStore

DeleteImportedKeyMaterialError
Errors returned by DeleteImportedKeyMaterial

DescribeCustomKeyStoresError
Errors returned by DescribeCustomKeyStores

DescribeKeyError
Errors returned by DescribeKey

DisableKeyError
Errors returned by DisableKey

DisableKeyRotationError
Errors returned by DisableKeyRotation

DisconnectCustomKeyStoreError
Errors returned by DisconnectCustomKeyStore

EnableKeyError
Errors returned by EnableKey

EnableKeyRotationError
Errors returned by EnableKeyRotation

EncryptError
Errors returned by Encrypt

GenerateDataKeyError
Errors returned by GenerateDataKey

GenerateDataKeyWithoutPlaintextError
Errors returned by GenerateDataKeyWithoutPlaintext

GenerateRandomError
Errors returned by GenerateRandom

GetKeyPolicyError
Errors returned by GetKeyPolicy

GetKeyRotationStatusError
Errors returned by GetKeyRotationStatus

GetParametersForImportError
Errors returned by GetParametersForImport

ImportKeyMaterialError
Errors returned by ImportKeyMaterial

ListAliasesError
Errors returned by ListAliases

ListGrantsError
Errors returned by ListGrants

ListKeyPoliciesError
Errors returned by ListKeyPolicies

ListKeysError
Errors returned by ListKeys

ListResourceTagsError
Errors returned by ListResourceTags

ListRetirableGrantsError
Errors returned by ListRetirableGrants

PutKeyPolicyError
Errors returned by PutKeyPolicy

ReEncryptError
Errors returned by ReEncrypt

RetireGrantError
Errors returned by RetireGrant

RevokeGrantError
Errors returned by RevokeGrant

ScheduleKeyDeletionError
Errors returned by ScheduleKeyDeletion

TagResourceError
Errors returned by TagResource

UntagResourceError
Errors returned by UntagResource

UpdateAliasError
Errors returned by UpdateAlias

UpdateCustomKeyStoreError
Errors returned by UpdateCustomKeyStore

UpdateKeyDescriptionError
Errors returned by UpdateKeyDescription

Traits

Kms
Trait representing the capabilities of the KMS API. KMS clients implement this trait.