Struct rusoto_kms::GenerateDataKeyWithoutPlaintextRequest

pub struct GenerateDataKeyWithoutPlaintextRequest {
    pub encryption_context: Option<HashMap<String, String>>,
    pub grant_tokens: Option<Vec<String>>,
    pub key_id: String,
    pub key_spec: Option<String>,
    pub number_of_bytes: Option<i64>,
}

Fields
encryption_context: Option<HashMap<String, String>>

A set of key-value pairs that represents additional authenticated data.

For more information, see Encryption Context in the AWS Key Management Service Developer Guide.

grant_tokens: Option<Vec<String>>

A list of grant tokens.

For more information, see Grant Tokens in the AWS Key Management Service Developer Guide.

key_id: String

The identifier of the CMK under which to generate and encrypt the data encryption key.

A valid identifier is the unique key ID or the Amazon Resource Name (ARN) of the CMK, or the alias name or ARN of an alias that refers to the CMK. Examples:

Unique key ID: 1234abcd-12ab-34cd-56ef-1234567890ab

CMK ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

Alias name: alias/ExampleAlias

Alias ARN: arn:aws:kms:us-east-2:111122223333:alias/ExampleAlias

key_spec: Option<String>

The length of the data encryption key. Use AES_128 to generate a 128-bit symmetric key, or AES_256 to generate a 256-bit symmetric key.

number_of_bytes: Option<i64>

The length of the data encryption key in bytes. For example, use the value 64 to generate a 512-bit data key (64 bytes is 512 bits). For common key lengths (128-bit and 256-bit symmetric keys), we recommend that you use the KeySpec field instead of this one.