Struct rusoto_kms::KeyMetadata

pub struct KeyMetadata {
    pub aws_account_id: Option<String>,
    pub arn: Option<String>,
    pub creation_date: Option<f64>,
    pub deletion_date: Option<f64>,
    pub description: Option<String>,
    pub enabled: Option<bool>,
    pub expiration_model: Option<String>,
    pub key_id: String,
    pub key_manager: Option<String>,
    pub key_state: Option<String>,
    pub key_usage: Option<String>,
    pub origin: Option<String>,
    pub valid_to: Option<f64>,
}

Contains metadata about a customer master key (CMK).

This data type is used as a response element for the CreateKey and DescribeKey operations.

Fields
aws_account_id: Option<String>

The twelve-digit account ID of the AWS account that owns the CMK.

arn: Option<String>

The Amazon Resource Name (ARN) of the CMK. For examples, see AWS Key Management Service (AWS KMS) in the Example ARNs section of the AWS General Reference.

creation_date: Option<f64>

The date and time when the CMK was created.

deletion_date: Option<f64>

The date and time after which AWS KMS deletes the CMK. This value is present only when KeyState is PendingDeletion, otherwise this value is omitted.

description: Option<String>

The description of the CMK.

enabled: Option<bool>

Specifies whether the CMK is enabled. When KeyState is Enabled this value is true, otherwise it is false.

expiration_model: Option<String>

Specifies whether the CMK's key material expires. This value is present only when Origin is EXTERNAL, otherwise this value is omitted.

key_id: String

The globally unique identifier for the CMK.

key_manager: Option<String>

The CMK's manager. CMKs are either customer-managed or AWS-managed. For more information about the difference, see Customer Master Keys in the AWS Key Management Service Developer Guide.

key_state: Option<String>

The state of the CMK.

For more information about how key state affects the use of a CMK, see How Key State Affects the Use of a Customer Master Key in the AWS Key Management Service Developer Guide.

key_usage: Option<String>

The cryptographic operations for which you can use the CMK. Currently the only allowed value is ENCRYPT_DECRYPT, which means you can use the CMK for the Encrypt and Decrypt operations.

origin: Option<String>

The source of the CMK's key material. When this value is AWS_KMS, AWS KMS created the key material. When this value is EXTERNAL, the key material was imported from your existing key management infrastructure or the CMK lacks key material.

valid_to: Option<f64>

The time at which the imported key material expires. When the key material expires, AWS KMS deletes the key material and the CMK becomes unusable. This value is present only for CMKs whose Origin is EXTERNAL and whose ExpirationModel is KEY_MATERIAL_EXPIRES, otherwise this value is omitted.