Crate rusoto_kms

AWS Key Management Service

If you're using the service, you're probably looking for KmsClient and Kms.

Structs

AliasListEntry
Contains information about an alias.

CancelKeyDeletionRequest

CancelKeyDeletionResponse

CreateAliasRequest

CreateGrantRequest

CreateGrantResponse

CreateKeyRequest

CreateKeyResponse

DecryptRequest

DecryptResponse

DeleteAliasRequest

DeleteImportedKeyMaterialRequest

DescribeKeyRequest

DescribeKeyResponse

DisableKeyRequest

DisableKeyRotationRequest

EnableKeyRequest

EnableKeyRotationRequest

EncryptRequest

EncryptResponse

GenerateDataKeyRequest

GenerateDataKeyResponse

GenerateDataKeyWithoutPlaintextRequest

GenerateDataKeyWithoutPlaintextResponse

GenerateRandomRequest

GenerateRandomResponse

GetKeyPolicyRequest

GetKeyPolicyResponse

GetKeyRotationStatusRequest

GetKeyRotationStatusResponse

GetParametersForImportRequest

GetParametersForImportResponse

GrantConstraints
A structure that you can use to allow certain operations in the grant only when the desired encryption context is present. For more information about encryption context, see Encryption Context in the AWS Key Management Service Developer Guide.

Grant constraints apply only to operations that accept encryption context as input. For example, the DescribeKey operation does not accept encryption context as input. A grant that allows the DescribeKey operation does so regardless of the grant constraints. In constrast, the Encrypt operation accepts encryption context as input. A grant that allows the Encrypt operation does so only when the encryption context of the Encrypt operation satisfies the grant constraints.

GrantListEntry
Contains information about an entry in a list of grants.

ImportKeyMaterialRequest

ImportKeyMaterialResponse

KeyListEntry
Contains information about each entry in the key list.

KeyMetadata
Contains metadata about a customer master key (CMK).

This data type is used as a response element for the CreateKey and DescribeKey operations.

KmsClient
A client for the KMS API.

ListAliasesRequest

ListAliasesResponse

ListGrantsRequest

ListGrantsResponse

ListKeyPoliciesRequest

ListKeyPoliciesResponse

ListKeysRequest

ListKeysResponse

ListResourceTagsRequest

ListResourceTagsResponse

ListRetirableGrantsRequest

PutKeyPolicyRequest

ReEncryptRequest

ReEncryptResponse

RetireGrantRequest

RevokeGrantRequest

ScheduleKeyDeletionRequest

ScheduleKeyDeletionResponse

Tag
A key-value pair. A tag consists of a tag key and a tag value. Tag keys and tag values are both required, but tag values can be empty (null) strings.

TagResourceRequest

UntagResourceRequest

UpdateAliasRequest

UpdateKeyDescriptionRequest

Enums

CancelKeyDeletionError
Errors returned by CancelKeyDeletion

CreateAliasError
Errors returned by CreateAlias

CreateGrantError
Errors returned by CreateGrant

CreateKeyError
Errors returned by CreateKey

DecryptError
Errors returned by Decrypt

DeleteAliasError
Errors returned by DeleteAlias

DeleteImportedKeyMaterialError
Errors returned by DeleteImportedKeyMaterial

DescribeKeyError
Errors returned by DescribeKey

DisableKeyError
Errors returned by DisableKey

DisableKeyRotationError
Errors returned by DisableKeyRotation

EnableKeyError
Errors returned by EnableKey

EnableKeyRotationError
Errors returned by EnableKeyRotation

EncryptError
Errors returned by Encrypt

GenerateDataKeyError
Errors returned by GenerateDataKey

GenerateDataKeyWithoutPlaintextError
Errors returned by GenerateDataKeyWithoutPlaintext

GenerateRandomError
Errors returned by GenerateRandom

GetKeyPolicyError
Errors returned by GetKeyPolicy

GetKeyRotationStatusError
Errors returned by GetKeyRotationStatus

GetParametersForImportError
Errors returned by GetParametersForImport

ImportKeyMaterialError
Errors returned by ImportKeyMaterial

ListAliasesError
Errors returned by ListAliases

ListGrantsError
Errors returned by ListGrants

ListKeyPoliciesError
Errors returned by ListKeyPolicies

ListKeysError
Errors returned by ListKeys

ListResourceTagsError
Errors returned by ListResourceTags

ListRetirableGrantsError
Errors returned by ListRetirableGrants

PutKeyPolicyError
Errors returned by PutKeyPolicy

ReEncryptError
Errors returned by ReEncrypt

RetireGrantError
Errors returned by RetireGrant

RevokeGrantError
Errors returned by RevokeGrant

ScheduleKeyDeletionError
Errors returned by ScheduleKeyDeletion

TagResourceError
Errors returned by TagResource

UntagResourceError
Errors returned by UntagResource

UpdateAliasError
Errors returned by UpdateAlias

UpdateKeyDescriptionError
Errors returned by UpdateKeyDescription

Traits

Kms
Trait representing the capabilities of the KMS API. KMS clients implement this trait.

Type Definitions

AWSAccountIdType

AlgorithmSpec

AliasList

AliasNameType

ArnType

BooleanType

CiphertextType

DataKeySpec

DateType

DescriptionType

EncryptionContextKey

EncryptionContextType

EncryptionContextValue

ErrorMessageType

ExpirationModelType

GrantIdType

GrantList

GrantNameType

GrantOperation

GrantOperationList

GrantTokenList

GrantTokenType

KeyIdType

KeyList

KeyState

KeyUsageType

LimitType

MarkerType

NumberOfBytesType

OriginType

PendingWindowInDaysType

PlaintextType

PolicyNameList

PolicyNameType

PolicyType

PrincipalIdType

TagKeyList

TagKeyType

TagList

TagValueType

WrappingKeySpec