Struct rusoto_kms::KmsClient

pub struct KmsClient { /* fields omitted */ }

A client for the KMS API.

Methods

impl KmsClient

pub fn new(region: Region) -> KmsClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> KmsClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,